% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_key.R
\name{get_api_key}
\alias{get_api_key}
\title{Get a Valid Alpha Vantage API Key}
\usage{
get_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{Character scalar or \code{NULL}. If \code{NULL}, fallback to environment
variable \code{ALPHA_VANTAGE_API_KEY}.}
}
\value{
Character scalar containing a non-empty API key.
}
\description{
Get a Valid Alpha Vantage API Key
}
