% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tickers.R
\name{validate_tickers}
\alias{validate_tickers}
\title{Validate Ticker Vector}
\usage{
validate_tickers(tickers)
}
\arguments{
\item{tickers}{Character vector of symbols.}
}
\value{
Invisibly returns \code{TRUE} when validation passes.
}
\description{
Validate Ticker Vector
}
